+++ integration-tests/src/test/java/org/zowe/apiml/integration/authentication/providers/SafLoginTest.java
@@ -11,16 +11,25 @@
 package org.zowe.apiml.integration.authentication.providers;
 
 import io.restassured.RestAssured;
+import org.apache.http.HttpHeaders;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.zowe.apiml.security.common.login.LoginRequest;
 import org.zowe.apiml.util.SecurityUtils;
 import org.zowe.apiml.util.TestWithStartedInstances;
 import org.zowe.apiml.util.categories.SAFAuthTest;
+import org.zowe.apiml.util.config.ConfigReader;
 
 import java.net.URI;
+import java.nio.charset.StandardCharsets;
+import java.util.Base64;
 
+import static io.restassured.RestAssured.given;
+import static io.restassured.http.ContentType.JSON;
+import static org.apache.http.HttpStatus.SC_UNAUTHORIZED;
+import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsNot.not;
@@ -52,4 +61,47 @@ class SafLoginTest implements TestWithStartedInstances {
             }
         }
     }
+
+    @Nested
+    class ExpiredPassword {
+
+        private final String USERNAME = ConfigReader.environmentConfiguration().getCredentials().getUser();
+        private final String EXPIRED_PASSWORD = "expiredPassword";
+
+        @ParameterizedTest(name = "givenExpiredAccountCredentialsInBody {index} {0} ")
+        @MethodSource("org.zowe.apiml.integration.authentication.providers.LoginTest#loginUrlsSource")
+        void givenExpiredAccountCredentialsInBody(URI loginUrl) {
+            LoginRequest loginRequest = new LoginRequest(USERNAME, EXPIRED_PASSWORD.toCharArray());
+
+            given()
+                .contentType(JSON)
+                .body(loginRequest)
+            .when()
+                .post(loginUrl)
+            .then()
+                .statusCode(is(SC_UNAUTHORIZED))
+                .body(
+                    "messages.find { it.messageNumber == 'ZWEAT412E' }.messageContent", containsString("expire")
+                );
+        }
+
+        @ParameterizedTest(name = "givenExpiredAccountCredentialsInHeader {index} {0} ")
+        @MethodSource("org.zowe.apiml.integration.authentication.providers.LoginTest#loginUrlsSource")
+        void givenExpiredAccountCredentialsInHeader(URI loginUrl) {
+            String headerValue = "Basic " + Base64.getEncoder().encodeToString((USERNAME + ":" + EXPIRED_PASSWORD).getBytes(StandardCharsets.UTF_8));
+
+            given()
+                .contentType(JSON)
+                .header(HttpHeaders.AUTHORIZATION, headerValue)
+            .when()
+                .post(loginUrl)
+            .then()
+                .statusCode(is(SC_UNAUTHORIZED))
+                .body(
+                    "messages.find { it.messageNumber == 'ZWEAT412E' }.messageContent", containsString("expire")
+                );
+        }
+
+    }
+
 }
